+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add `Enable llms.txt` and `Enable llms-full.txt` site settings to disable the aggregated files per site while keeping the markdown pages available. ([#24](https://github.com/samuelreichor/craft-llmify/issues/24))
+- Add an `Include Social Links` site setting with an editable `Social Links` table that adds a `## Social` section to `llms.txt` and `llms-full.txt`. The table follows SEOmatic's "Same As URLs" until changes are saved that differ from it; social links can also be maintained without SEOmatic. ([#24](https://github.com/samuelreichor/craft-llmify/issues/24))
 
 ## 1.9.0 - 2026-07-18